+++ Source/Sitecore.FakeDb/Pipelines/IPipelineProcessor.cs
@@ -1,0 +1,9 @@
+﻿namespace Sitecore.FakeDb.Pipelines
+{
+  using Sitecore.Pipelines;
+
+  public interface IPipelineProcessor
+  {
+    void Process(PipelineArgs args);
+  }
+}