+++ Source/Sitecore.FakeDb/Pipelines/PipelineWatcherProcessor.cs
@@ -1,7 +1,11 @@
 namespace Sitecore.FakeDb.Pipelines
 {
-  using Sitecore.Pipelines;
   using System;
+  using Sitecore.Data;
+  using Sitecore.Diagnostics;
+  using Sitecore.FakeDb.Data.DataProviders;
+  using Sitecore.FakeDb.Data.Engines;
+  using Sitecore.Pipelines;
 
   public class PipelineWatcherProcessor
   {
@@ -11,9 +15,15 @@ namespace Sitecore.FakeDb.Pipelines
 
     public PipelineWatcherProcessor(string pipelineName)
     {
-      Diagnostics.Assert.ArgumentNotNullOrEmpty(pipelineName, "pipelineName");
+      Assert.ArgumentNotNullOrEmpty(pipelineName, "pipelineName");
 
       this.pipelineName = pipelineName;
+
+      // TODO:[High] Open question: how to pass the DataStorage to this class?
+      if (this.DataStorage == null)
+      {
+        this.DataStorage = ((FakeDataProvider)Database.GetDatabase("master").GetDataProviders()[0]).DataStorage;
+      }
     }
 
     public string PipelineName
@@ -21,11 +31,20 @@ namespace Sitecore.FakeDb.Pipelines
       get { return this.pipelineName; }
     }
 
+    public DataStorage DataStorage { get; set; }
+
     public virtual void Process(PipelineArgs args)
     {
+      Assert.ArgumentNotNull(args, "args");
+
       if (PipelineRun != null)
       {
         PipelineRun(this, new PipelineRunEventArgs(this.PipelineName, args));
+      }
+
+      if (this.DataStorage != null && this.DataStorage.Pipelines.ContainsKey(this.PipelineName))
+      {
+        this.DataStorage.Pipelines[this.PipelineName].Process(args);
       }
     }
   }