+++ Source/Sitecore.FakeDb/Properties/AssemblyVersionInfo.cs
@@ -1,5 +1,5 @@
 ﻿using System.Reflection;
 
-[assembly: AssemblyVersion("0.19.1.0")]
-[assembly: AssemblyFileVersion("0.19.1.0")]
-[assembly: AssemblyInformationalVersion("0.19.1")]
+[assembly: AssemblyVersion("0.20.0.0")]
+[assembly: AssemblyFileVersion("0.20.0.0")]
+[assembly: AssemblyInformationalVersion("0.20.0")]